pyproject.toml
[tool.poetry]
name = "appsec"
version = "0.1.0"
description = ""
authors = ["Aaron Joyce <joyceaa@tcd.ie>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
asyncio = "^3.4.3"
aiohttp = "^3.10.2"
rich = "^13.8.1"
autogen-agentchat = ">=0.2,<1.0"
pydantic = "^2.9.2"
langchain = "^0.3.1"
pygithub = "^2.4.0"
langchain-anthropic = "^0.2.1"
langchain-openai = "^0.2.1"
dynaconf = "^3.2.6"
python-dotenv = "^1.0.1"
autogen = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
appsec = "appsec.cli:cli"
